perform input operation requires a valid sessionID - working on this bug
