+++ MetaCocktailsSwiftData/Views/CreateACocktailView/AddIngredientView.swift
@@ -9,14 +9,14 @@ import SwiftUI
 
 struct AddIngredientView: View {
     
-    @Bindable var viewModel = CreateACocktailViewModel()
-    @Environment(\.dismiss) private var dismiss
+    @Bindable var viewModel = AddCocktailViewModel()
+    @Binding var isShowingAddIngredients: Bool
     @FocusState private var keyboardFocused: Bool
     
     var body: some View {
         VStack {
             SearchBarView(searchText: $viewModel.ingredientName)
-                .focused($keyboardFocused)
+                //.focused($keyboardFocused)
                 .onChange(of: viewModel.ingredientName) {
                     viewModel.matchAllPhysicalCocktailComponents()
                 }
@@ -30,12 +30,14 @@ struct AddIngredientView: View {
                                     viewModel.ingredientType = category
                                 }
                                 viewModel.dynamicallyChangeMeasurementUnit()
+                                viewModel.ingredientName = component.name
                             }
                             if component.isSpirit {
                                 if let category = component.spiritCategory {
                                     viewModel.ingredientType = category
                                 }
                                 viewModel.dynamicallyChangeMeasurementUnit()
+                                viewModel.ingredientName = component.name
                             }
                             
                             
@@ -72,15 +74,15 @@ struct AddIngredientView: View {
                     viewModel.addedIngredients.append(CocktailIngredient(viewModel.currentSelectedComponent.nACategory ?? IngredientType.juices(.carrotJuice), value: viewModel.ingredientAmount, unit: viewModel.selectedMeasurementUnit))
                 }
                 for ingredients in viewModel.addedIngredients {
-                    print("\(ingredients.value) \(ingredients.unit)\(ingredients.ingredient.name)")
+                    print("\(ingredients.value) \(ingredients.unit) \(ingredients.ingredient.name)")
                 }
                 viewModel.ingredientName = ""
                 viewModel.ingredientType = IngredientType.agaves(.elTesoroRepo)
                 viewModel.ingredientAmount = 0
                 viewModel.selectedMeasurementUnit = .fluidOunces
                 
-                
-                dismiss()
+                isShowingAddIngredients.toggle()
+               
             } label: {
                 Text("Add to Ingredients")
                 
@@ -97,5 +99,5 @@ struct AddIngredientView: View {
 }
 
 #Preview {
-    AddIngredientView()
+    AddIngredientView(isShowingAddIngredients: .constant(true))
 }
